AG-13591 - Refactor out property modules into separate component
