+++ documentation/ag-grid-docs/src/components/reference-documentation/components/PropertyModules.tsx
@@ -1,0 +1,102 @@
+import type { Framework } from '@ag-grid-types';
+import { Icon } from '@ag-website-shared/components/icon/Icon';
+import styles from '@ag-website-shared/components/reference-documentation/ApiReference.module.scss';
+import { urlWithPrefix } from '@utils/urlWithPrefix';
+import classnames from 'classnames';
+import { type FunctionComponent, useCallback, useEffect, useState } from 'react';
+
+export const PropertyModules: FunctionComponent<{
+    modules: Array;
+    framework: Framework;
+}> = ({ modules, framework }) => {
+    const firstModule = modules[0];
+    const otherModules = modules.slice(1) ?? [];
+
+    const [isModuleTooltipVisible, setIsModuleTooltipVisible] = useState(false);
+
+    const toggleModuleTooltip = useCallback(() => {
+        setIsModuleTooltipVisible((prev) => !prev);
+    }, []);
+
+    // Close tooltip when clicking outside
+    useEffect(() => {
+        const handleClickOutside = (event: MouseEvent) => {
+            if (isModuleTooltipVisible) {
+                const target = event.target as HTMLElement;
+                if (!target.closest(`.${styles.moduleItem}`)) {
+                    setIsModuleTooltipVisible(false);
+                }
+            }
+        };
+
+        document.addEventListener('click', handleClickOutside);
+        return () => {
+            document.removeEventListener('click', handleClickOutside);
+        };
+    }, [isModuleTooltipVisible]);
+
+    return (
+        <div className={classnames(styles.metaItem, styles.moduleItem)}>
+            <div className={styles.moduleContent}>
+                <a
+                    href={urlWithPrefix({
+                        url: './modules',
+                        framework,
+                    })}
+                >
+                    <Icon name="module" />
+                    <span>{firstModule.name}</span>
+                    {firstModule.isEnterprise && (
+                        <span className={styles.enterpriseIcon}>
+                            <Icon name="enterprise" />
+                        </span>
+                    )}
+                </a>
+
+                {otherModules.length > 0 && (
+                    <>
+                        <span
+                            className={classnames(styles.moduleCount, {
+                                [styles.moduleCountActive]: isModuleTooltipVisible,
+                            })}
+                            onClick={(e) => {
+                                e.preventDefault();
+                                e.stopPropagation();
+                                toggleModuleTooltip();
+                            }}
+                        >
+                            +{otherModules.length} <Icon name="chevronDown" />
+                        </span>
+
+                        <div
+                            className={classnames(styles.moduleTooltip, {
+                                [styles.isVisible]: isModuleTooltipVisible,
+                            })}
+                        >
+                            <div className={styles.moduleTooltipContent}>
+                                {otherModules.map((module) => (
+                                    <div key={module.name}>
+                                        <a
+                                            href={urlWithPrefix({
+                                                url: './modules/',
+                                                framework,
+                                            })}
+                                        >
+                                            <Icon name="module" />
+                                            {module.name}
+                                            {module.isEnterprise && (
+                                                <span className={styles.enterpriseIcon}>
+                                                    <Icon name="enterprise" />
+                                                </span>
+                                            )}
+                                        </a>
+                                    </div>
+                                ))}
+                            </div>
+                        </div>
+                    </>
+                )}
+            </div>
+        </div>
+    );
+};
